Consolidate package-name parsing and extract stack frame formatter
Move the scoped-package fallback parser into parse-package-name as a single
resolvePackageName, so source-frame-policy no longer owns parsing. Extract
formatStackFrameLine from formatStackContext so the loop is compute/dedup/push.
