+++ packages/react-grab/src/utils/parse-package-name.ts
@@ -96,3 +96,47 @@ export const parsePackageName = (fileName: string | null | undefined): string | 
 
   return null;
 };
+
+const SCOPED_PACKAGE_PATTERN = /^@[A-Za-z0-9][A-Za-z0-9._-]*$/;
+const PACKAGE_NAME_SEGMENT_PATTERN = /^[A-Za-z0-9][A-Za-z0-9._-]*$/;
+// A relative scoped path like `../@acme/app/...` has no node_modules marker to
+// prove it is third-party, so a monorepo's own app workspace would otherwise be
+// misread as a package. Best-effort allowlist of common first-party app dirs.
+const APPLICATION_PACKAGE_NAME_SEGMENTS = new Set(["app", "web", "website", "frontend", "client"]);
+
+const stripRelativeSourcePathPrefix = (path: string): string | null => {
+  let remainingPath = path;
+  let didStripPrefix = false;
+  while (remainingPath.startsWith("../") || remainingPath.startsWith("./")) {
+    didStripPrefix = true;
+    remainingPath = remainingPath.slice(remainingPath.startsWith("../") ? 3 : 2);
+  }
+  return didStripPrefix ? remainingPath : null;
+};
+
+const parseScopedPackageSourceName = (fileName: string): string | null => {
+  const sourcePath = stripRelativeSourcePathPrefix(
+    safeDecodeURIComponent(normalizeFileName(fileName)),
+  );
+  if (!sourcePath) return null;
+
+  const [scope, packageName] = splitPathSegments(sourcePath);
+  if (
+    !scope ||
+    !packageName ||
+    !SCOPED_PACKAGE_PATTERN.test(scope) ||
+    !PACKAGE_NAME_SEGMENT_PATTERN.test(packageName) ||
+    APPLICATION_PACKAGE_NAME_SEGMENTS.has(packageName)
+  ) {
+    return null;
+  }
+
+  return `${scope}/${packageName}`;
+};
+
+// parsePackageName keys off node_modules/.vite/CDN markers; the scoped fallback
+// handles bare relative imports like `../@acme/ui/...` that carry no marker.
+export const resolvePackageName = (fileName: string | null | undefined): string | null => {
+  if (!fileName) return null;
+  return parsePackageName(fileName) ?? parseScopedPackageSourceName(fileName);
+};
