+++ packages/react-grab/src/utils/source-frame-policy.ts
@@ -1,20 +1,13 @@
-import { isSourceFile, normalizeFileName } from "bippy/source";
+import { isSourceFile } from "bippy/source";
 import type { SourceOptions } from "../types.js";
 import { normalizeFilePath } from "./normalize-file-path.js";
-import { parsePackageName } from "./parse-package-name.js";
-import { safeDecodeURIComponent } from "./safe-decode-uri-component.js";
+import { resolvePackageName } from "./parse-package-name.js";
 
 // Always ignored, in addition to any user-supplied ignorePaths: design-system
 // wrappers (e.g. shadcn's components/ui) are rarely the file a user wants to
 // edit, so grabs resolve to the consuming app source instead.
 const DEFAULT_IGNORED_SOURCE_PATHS: readonly string[] = ["components/ui"];
 const PATH_SEPARATOR_PATTERN = /[/\\]/;
-const SCOPED_PACKAGE_PATTERN = /^@[A-Za-z0-9][A-Za-z0-9._-]*$/;
-const PACKAGE_NAME_SEGMENT_PATTERN = /^[A-Za-z0-9][A-Za-z0-9._-]*$/;
-// A relative scoped path like `../@acme/app/...` has no node_modules marker to
-// prove it is third-party, so a monorepo's own app workspace would otherwise be
-// misread as a package. Best-effort allowlist of common first-party app dirs.
-const APPLICATION_PACKAGE_NAME_SEGMENTS = new Set(["app", "web", "website", "frontend", "client"]);
 // Keyed by fileName and only used when no custom ignorePaths are set; valid only
 // while default classification depends solely on fileName.
 const defaultClassificationCache = new Map<string, SourcePathClassification>();
@@ -72,36 +65,6 @@ const matchesIgnoredSourcePath = (fileName: string, sourceOptions?: SourceOption
   return false;
 };
 
-const stripRelativeSourcePathPrefix = (path: string): string | null => {
-  let remainingPath = path;
-  let didStripPrefix = false;
-  while (remainingPath.startsWith("../") || remainingPath.startsWith("./")) {
-    didStripPrefix = true;
-    remainingPath = remainingPath.slice(remainingPath.startsWith("../") ? 3 : 2);
-  }
-  return didStripPrefix ? remainingPath : null;
-};
-
-const parseScopedPackageSourceName = (fileName: string): string | null => {
-  const sourcePath = stripRelativeSourcePathPrefix(
-    safeDecodeURIComponent(normalizeFileName(fileName)),
-  );
-  if (!sourcePath) return null;
-
-  const [scope, packageName] = sourcePath.split(PATH_SEPARATOR_PATTERN).filter(Boolean);
-  if (
-    !scope ||
-    !packageName ||
-    !SCOPED_PACKAGE_PATTERN.test(scope) ||
-    !PACKAGE_NAME_SEGMENT_PATTERN.test(packageName) ||
-    APPLICATION_PACKAGE_NAME_SEGMENTS.has(packageName)
-  ) {
-    return null;
-  }
-
-  return `${scope}/${packageName}`;
-};
-
 export const classifySourcePath = (
   fileName: string | null | undefined,
   sourceOptions?: SourceOptions,
@@ -126,7 +89,7 @@ const classifySourcePathUncached = (
     return { kind: "unknown", packageName: null };
   }
 
-  const packageName = parsePackageName(fileName) ?? parseScopedPackageSourceName(fileName);
+  const packageName = resolvePackageName(fileName);
   if (packageName) {
     return { kind: "package-source", packageName };
   }
